# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minha API",
        "description": "Documentação da API",
        "contact": {
            "email": "seu-email@dominio.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal"
        }
    ],
    "paths": {
        "/api/client": {
            "get": {
                "tags": [
                    "public"
                ],
                "summary": "Obter lista de Clients",
                "description": "Retorna uma lista",
                "operationId": "ClientList",
                "responses": {
                    "200": {
                        "description": "Lista de exemplos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "public"
                ],
                "summary": "Inseri um objeto",
                "description": "Cria um novo recurso no banco de dados",
                "operationId": "ClientStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "document",
                                    "type_document"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "andre"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "an@gm.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "231321321"
                                    },
                                    "document": {
                                        "type": "string",
                                        "example": "231321321"
                                    },
                                    "type_document": {
                                        "type": "string",
                                        "example": "cpf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exemplo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/client/{id}": {
            "put": {
                "tags": [
                    "public"
                ],
                "summary": "Atualizar um objeto",
                "description": "Atualiza um recurso existente com base no ID",
                "operationId": "ClientUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "document",
                                    "type_document"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "andre"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "an@gm.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "231321321"
                                    },
                                    "document": {
                                        "type": "string",
                                        "example": "231321321"
                                    },
                                    "type_document": {
                                        "type": "string",
                                        "example": "cpf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exemplo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exemplo não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "public"
                ],
                "summary": "Deletar um recurso",
                "description": "Deleta um recurso baseado no ID fornecido",
                "operationId": "ClientDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do recurso que será deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso deletado com sucesso"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                }
            }
        },
        "/api/equipament": {
            "get": {
                "tags": [
                    "public"
                ],
                "summary": "Retornar uma lista de equipamentos",
                "description": "Retornar uma lista de equipamentos",
                "operationId": "equipamentList",
                "responses": {
                    "200": {
                        "description": "Lista de exemplos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Equipament"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "public"
                ],
                "summary": "Inseri um objeto",
                "description": "Cria um novo recurso no banco de dados",
                "operationId": "EquipamentStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "uf": {
                                        "type": "string",
                                        "example": "Mola"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exemplo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Equipament"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/equipamento/{id}": {
            "put": {
                "tags": [
                    "public"
                ],
                "summary": "Atualizar um objeto",
                "description": "Atualiza um recurso existente com base no ID",
                "operationId": "EquipamentUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Equipamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "uf"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Novo Nome do Exemplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exemplo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Equipament"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exemplo não encontrado"
                    }
                }
            }
        },
        "/api/equipament/{id}": {
            "delete": {
                "tags": [
                    "public"
                ],
                "summary": "Deletar um recurso",
                "description": "Deleta um recurso baseado no ID fornecido",
                "operationId": "EquipamentDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do recurso que será deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso deletado com sucesso"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Equipament": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Mola"
                    }
                },
                "type": "object"
            },
            "Client": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Andre"
                    },
                    "email": {
                        "type": "string",
                        "example": "andre@gmail.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "21312321"
                    },
                    "document": {
                        "type": "string",
                        "example": "4490837"
                    },
                    "type_document": {
                        "type": "string",
                        "example": "cpf"
                    }
                },
                "type": "object"
            }
        }
    },
...
}